Migrate policy.Target
