+++ packages/services/api/src/modules/policy/resolvers.ts
@@ -5,37 +5,9 @@ import { IdTranslator } from '../shared/providers/id-translator';
 import { PolicyModule } from './__generated__/types';
 import { SchemaPolicyApiProvider } from './providers/schema-policy-api.provider';
 import { SchemaPolicyProvider } from './providers/schema-policy.provider';
-import { formatTRPCErrors, policyInputToConfigObject, serializeSeverity } from './utils';
+import { formatTRPCErrors, policyInputToConfigObject } from './utils';
 
 export const resolvers: PolicyModule.Resolvers = {
-  Target: {
-    schemaPolicy: async (target, _, { injector }) => {
-      const { mergedPolicy, orgLevel, projectLevel } = await injector
-        .get(SchemaPolicyProvider)
-        .getCalculatedTargetPolicyForApi({
-          project: target.projectId,
-          organization: target.orgId,
-          target: target.id,
-        });
-
-      if (!mergedPolicy) {
-        return null;
-      }
-
-      const availableRules = await injector.get(SchemaPolicyApiProvider).listAvailableRules();
-      const rules = Object.entries(mergedPolicy).map(([ruleId, config]) => ({
-        rule: availableRules.find(r => r.name === ruleId)!,
-        severity: serializeSeverity(config[0]),
-        configuration: config[1],
-      }));
-
-      return {
-        mergedRules: rules,
-        organizationPolicy: orgLevel,
-        projectPolicy: projectLevel,
-      };
-    },
-  },
   SchemaPolicyRule: {
     id: r => r.name,
     description: r => r.description,
